src/main/java/org/codeus/Trie.java
package org.codeus;

import org.codeus.util.ExerciseNotCompletedException;

/**
 * Represents a Trie (prefix tree or k-ary tree) data structure for storing and managing strings.
 * A Trie can be highly effective for scenarios that involve fast prefix-based search operations,
 * such as autocomplete and spell checking.
 * <br>
 * Time Complexity:
 * <ul>
 * <li> Insertion: O(m), where m is the length of the word being inserted.</li>
 * <li> Search (contains): O(m), where m is the length of the word being searched.</li>
 * <li> Prefix Search (containsPrefix): O(m), where m is the length of the prefix.</li>
 * <li> Deletion: O(m), where m is the length of the word being removed.</li>
 * </ul>
 *
 * In the worst case, the space complexity is O(n * m), where n is the number of words stored in the Trie
 * and m is the average length of these words, although space efficiency is improved by sharing common prefixes.
 */
public class Trie {

  public static final int ALPHABET_SIZE = 26;

  /**
   * Constructs a new Trie instance.
   * Initializes a {@link Trie#rootNode} with the ' ' character.
   */
  public Trie() {
    throw new ExerciseNotCompletedException();
  }

  /**
   * Returns the index of a given character in the English alphabet for internal Trie node positioning.
   *
   * @param c the character for which to determine the index
   * @return the index of the character within the alphabet (0-based, starting at 'a')
   */
  private int getBucketIndex(char c) {
    throw new ExerciseNotCompletedException();
  }

  /**
   * Inserts a string into the Trie.
   * The input string is transformed to lowercase format before insertion.
   * This implementation is duplicate agnostic, meaning if you insert the same value twice,
   * it should not affect the internal structure of the Trie.
   * This method validates the input using {@link #validateInput(String)}, which throws an
   * IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters.
   *
   *
   * @param value the string to insert into the Trie
   * @throws IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters
   */
  public void insert(String value) {
    throw new ExerciseNotCompletedException();
  }


  /**
   * Checks if a given string is present in the Trie as a complete word.
   * The input string is transformed to lowercase format before checking.
   * This method validates the input using {@link #validateInput(String)}, which throws an
   * IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters.
   *
   * @param value the string to check for in the Trie
   * @return true if the string is present as a complete word, false otherwise
   * @throws IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters
   */
  public boolean contains(String value) {
    throw new ExerciseNotCompletedException();
  }


  /**
   * Checks if a given string is present in the Trie as a prefix.
   * The input string is transformed to lowercase format before checking.
   * This method validates the input using {@link #validateInput(String)}, which throws an
   * IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters.
   *
   * @param value the prefix string to check for in the Trie
   * @return true if the string is present as a prefix, false otherwise
   * @throws IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters
   */
  public boolean containsPrefix(String value) {
    throw new ExerciseNotCompletedException();
  }

  /**
   * Removes a string from the Trie if it exists as a complete word.
   * The input string is transformed to lowercase format before removal.
   * This method validates the input using {@link #validateInput(String)}, which throws an
   * IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters.
   * <br>
   * <p>The removal operation considers three possible cases:</p>
   * <ul>
   * <li><b>1. Isolated Word</b>: If the word exists as a complete word with no other word branching from
   *    its nodes, all nodes for the word are removed from the Trie.</li>
   * <li><b>2. Prefix of Another Word</b>: If the word to be removed is a prefix of another word
   *    (e.g., removing "cat" when "cater" exists), only the end-of-word marker for the word is removed,
   *    while the nodes are retained.</li>
   * <li><b>3. Contains Another Word as a Prefix</b>: If the word contains another word as a prefix
   *    (e.g., removing "cater" when "cat" exists), only the nodes beyond the shorter word are removed,
   *    leaving "cat" intact.</li>
   * </ul>
   *
   * @param value the string to remove from the Trie
   * @throws IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters
   */
  public void remove(String value) {
    throw new ExerciseNotCompletedException();
  }

  /**
   * Represents a node within the Trie, holding a character and links to child nodes.
   */
  public static class Node {

    /**
     * Constructs a new Trie Node with a specified character value.<br>
     * Initializes {@link Node#value} with the specified value, an array for {@link Node#children} nodes corresponding to the {@link Trie#ALPHABET_SIZE}
     * and {@link Node#isWordEnd} with initial `false` value.
     *
     * @param value the character value associated with this node
     */
    public Node(char value) {
      throw new ExerciseNotCompletedException();
    }

    /**
     * Counts the non-null children of this node.
     *
     * @return the number of non-null children
     */
    int getChildrenCount() {
      throw new ExerciseNotCompletedException();
    }

    /**
     * Checks if this node has any non-null children.
     *
     * @return true if the node has no children, false otherwise
     */
    boolean isEmpty() {
      throw new ExerciseNotCompletedException();
    }
  }

  /**
   * Validates the input string for Trie operations.
   * Throws an IllegalArgumentException if the input is null, empty, or contains
   * non-alphabetic characters.
   *
   * @param input the string to validate
   * @throws IllegalArgumentException if the input is null, empty, or contains non-alphabetic characters
   */
  private void validateInput(String input) {
    if (input == null || input.trim().isEmpty()) {
      throw new IllegalArgumentException("Input must not be null or empty.");
    }
    // Match only alphabetic characters (Unicode supported)
    if (!input.matches("[\\p{L}]+")) {
      throw new IllegalArgumentException("Input must contain only alphabetic characters.");
    }
  }

}
